# Remove explicit stream finish