--- a/quic-client/src/nonblocking/quic_client.rs
+++ b/quic-client/src/nonblocking/quic_client.rs
@@ -274,7 +274,7 @@ impl QuicClient {
         let mut send_stream = connection.open_uni().await?;
 
         send_stream.write_all(data).await?;
-        send_stream.finish().await?;
+        // send_stream.finish().await?;
         Ok(())
     }