#### Problem

stream finish is redundant as StreamDrop is already doing that.
Removing that will allow next call to be made faster without waiting for the stream to be finished.

#### Summary of Changes

removed explicit stream finish

Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->